arping: Fix comparison of different signedness warning
../arping.c:768:45: warning: comparison of integer expressions of different signedness: ‘int’ and ‘uint64_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  768 |         (ctl->quit_on_reply && ctl->timeout < total_expires)) {

Fixes: 1df5350 ("arping: fix -f quit on first reply regression")

+++ arping.c
@@ -765,7 +765,7 @@ static int event_loop(struct run_state *ctl)
 				}
 				total_expires += exp;
 				if ((0 < ctl->count && (uint64_t)ctl->count < total_expires) ||
-				    (ctl->quit_on_reply && ctl->timeout < total_expires)) {
+				    (ctl->quit_on_reply && ctl->timeout < (int)total_expires)) {
 					exit_loop = 1;
 					continue;
 				}
